requirements.txt
chumpy
easydict
ipdb
kornia==0.6.12
loguru
matplotlib
numpy==1.23.1
open3d
opencv-contrib-python-headless==4.8.0.74
opencv-python==4.8.0.74
pillow==10.4.0
pygit2==1.14.1
pymeshlab==2022.2.post4
pyparsing
pytorch-lightning==2.3.0
scipy==1.10.1
scikit-learn
tqdm
trimesh
omegaconf
plyfile
moderngl-window==2.4.6
huggingface_hub==0.23.4


# threestudio
jaxtyping
typeguard
git+https://github.com/KAIR-BAIR/nerfacc.git@v0.5.2
git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch
diffusers<0.20
transformers==4.41.2
tensorboard
imageio>=2.28.0
imageio[ffmpeg]
git+https://github.com/NVlabs/nvdiffrast.git

libigl
xatlas
trimesh[easy]
networkx
pysdf
PyMCubes
wandb
gradio==4.11.0
git+https://github.com/ashawkey/envlight.git
torchmetrics

# for zero123
einops
kornia
taming-transformers-rom1504
git+https://github.com/openai/CLIP.git
